+++ backend/canisters/user/api/src/queries/initial_state.rs
@@ -3,8 +3,8 @@ use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
 use types::{
-    CanisterId, Chat, ChatId, ChitEarned, DirectChatSummary, Empty, GroupChatSummary, PinNumberSettings, TimestampMillis,
-    UserId,
+    CanisterId, Chat, ChatId, ChitEarned, DirectChatSummary, Empty, GroupChatSummary, PinNumberSettings, StreakInsurance,
+    TimestampMillis, UserId,
 };
 
 pub type Args = Empty;
@@ -34,6 +34,7 @@ pub struct SuccessResult {
     pub chit_balance: i32,
     pub streak: u16,
     pub streak_ends: TimestampMillis,
+    pub streak_insurance: Option<StreakInsurance>,
     pub next_daily_claim: TimestampMillis,
     pub is_unique_person: bool,
     pub wallet_config: WalletConfig,